V22F-493 Tick to ulong

+++ speed-dreams/source-2.2.3/src/simulatedDrivingAssistance/backendUnitTests/mocks/BlackBoxDataMock.h
@@ -6,7 +6,7 @@
 struct BlackBoxDataMock
 {
 public:
-    BlackBoxDataMock(tCarElt* p_car, tSituation* p_situation, int p_tickCount, tTrackSeg* p_trackSegments, int p_trackSegmentsCount)
+    BlackBoxDataMock(tCarElt* p_car, tSituation* p_situation, unsigned long p_tickCount, tTrackSeg* p_trackSegments, int p_trackSegmentsCount)
     {
         Car = *p_car;
         Situation = *p_situation;
@@ -14,14 +14,14 @@ public:
     }
     tCarElt Car;
     tSituation Situation;
-    int TickCount;
+    unsigned long TickCount;
 };
 
 inline BlackBoxDataMock GetExampleBlackBoxDataMock()
 {
     tCarElt* car = new tCarElt();
     tSituation* situation = new tSituation();
-    int tickCount = 0;
+    unsigned long tickCount = 0;
 
     return { car, situation, tickCount, nullptr, 0 };
 }
@@ -30,7 +30,7 @@ inline BlackBoxDataMock CreateRandomBlackBoxDataMock(Random& random)
 {
     tCarElt* car = new tCarElt();
     tSituation* situation = new tSituation();
-    int tickCount = random.NextInt(0,1000);
+    unsigned long tickCount = random.NextInt(0,1000);
 
     return { car, situation, tickCount, nullptr, 0 };
 }
